refactor(#214): simplify prose decoding to per-node dispatch
